+++ apps/frontend/src/components/global/HelpAboutDropdown.vue
